LPD-46991 - Refactor. Move loadData to its own file (TS) and add miniumn test coverage.
Move utilities used only by loadData from utils/index to loadData
