+++ modules/apps/frontend-data-set/frontend-data-set-web/package.json
@@ -24,6 +24,7 @@
 		"classnames": "2.3.1",
 		"frontend-js-components-web": "*",
 		"frontend-js-web": "*",
+		"jest-fetch-mock": "3.0.3",
 		"prop-types": "15.7.2",
 		"react": "18.2.0"
 	},
