+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/utils/index.js
@@ -7,7 +7,6 @@ import {fetch} from 'frontend-js-web';
 
 import {DEFAULT_FETCH_HEADERS} from '../constants';
 import getValueFromItem from './getValueFromItem';
-import createOdataFilter from './odata';
 
 export function getData(apiURL, query) {
 	const url = new URL(apiURL);
@@ -86,185 +85,6 @@ export function createSortingString(values) {
 			}:${value.direction}`;
 		})
 		.join(',');
-}
-
-export function getFiltersString(odataFiltersStrings, providedFilters) {
-	let filtersString = '';
-
-	if (providedFilters) {
-		filtersString += providedFilters;
-	}
-
-	if (providedFilters && odataFiltersStrings.length) {
-		filtersString += ' and ';
-	}
-
-	if (odataFiltersStrings.length) {
-		filtersString += createOdataFilter(odataFiltersStrings);
-	}
-
-	return filtersString;
-}
-
-export async function loadData(
-	apiURL,
-	currentURL,
-	odataFiltersStrings,
-	searchParam,
-	delta,
-	page = 1,
-	sorts = [],
-	additionalAPIURLParameters
-) {
-	const fullUrl = apiURL.startsWith('/')
-		? Liferay.ThemeDisplay.getPortalURL() +
-			Liferay.ThemeDisplay.getPathContext() +
-			apiURL
-		: apiURL;
-
-	const url = new URL(fullUrl);
-
-	if (currentURL) {
-		url.searchParams.set('currentURL', currentURL);
-	}
-
-	const providedFilters = url.searchParams.get('filter');
-
-	url.searchParams.delete('filter');
-
-	if (providedFilters || odataFiltersStrings.length) {
-		url.searchParams.append(
-			'filter',
-			getFiltersString(odataFiltersStrings, providedFilters)
-		);
-	}
-
-	if (Liferay.ThemeDisplay.isImpersonated()) {
-		url.searchParams.append('doAsUserId', Liferay.ThemeDisplay.getUserId());
-	}
-
-	url.searchParams.append('page', page);
-	url.searchParams.append('pageSize', delta);
-
-	if (searchParam) {
-		url.searchParams.append('search', searchParam);
-	}
-
-	if (sorts.length) {
-		const updatedSorts = sorts.map((sort) => {
-			const key = sort.key.includes('LANG')
-				? sort.key.split(',')[0]
-				: sort.key;
-
-			return {
-				...sort,
-				key,
-			};
-		});
-
-		url.searchParams.set(
-			'sort',
-			updatedSorts
-				.map((item) => `${item.key}:${item.direction}`)
-				.join(',')
-		);
-	}
-
-	if (additionalAPIURLParameters) {
-		const additionalAPIURLParametersArray =
-			additionalAPIURLParameters.split('&');
-
-		additionalAPIURLParametersArray.forEach((parameter) => {
-			const [key, value] = parameter.split('=');
-
-			const existingFilter = url.searchParams.get('filter');
-
-			if (key === 'filter' && existingFilter) {
-				url.searchParams.set(
-					'filter',
-					`(${existingFilter}) and (${value})`
-				);
-			}
-			else if (key === 'sort' && url.searchParams.get('sort')) {
-				const newSortParams = [];
-
-				const existingSortArray = url.searchParams
-					.get('sort')
-					.split(',');
-
-				const existingSortParamFields = existingSortArray.map(
-					(sort) => sort.split(':')[0]
-				);
-
-				const additionalAPIURLParametersSortValueArray =
-					value.split(',');
-
-				additionalAPIURLParametersSortValueArray.forEach(
-					(additionalAPIURLParametersSortValueItem) => {
-						if (
-							!existingSortParamFields.includes(
-								additionalAPIURLParametersSortValueItem.split(
-									':'
-								)[0]
-							)
-						) {
-							newSortParams.push(
-								additionalAPIURLParametersSortValueItem
-							);
-						}
-					}
-				);
-
-				url.searchParams.set(
-					'sort',
-					existingSortArray.concat(newSortParams).join(',')
-				);
-			}
-			else if (
-				key === 'nestedFields' &&
-				url.searchParams.get('nestedFields')
-			) {
-				const existingNestedFieldsArray = url.searchParams
-					.get('nestedFields')
-					.split(',');
-				const additionalAPIURLParametersNestedFieldsValueArray =
-					value.split(',');
-
-				const newNestedFields = [...existingNestedFieldsArray];
-
-				additionalAPIURLParametersNestedFieldsValueArray.forEach(
-					(additionalAPIURLParametersNestedFieldsItem) => {
-						if (
-							!existingNestedFieldsArray.includes(
-								additionalAPIURLParametersNestedFieldsItem
-							)
-						) {
-							newNestedFields.push(
-								additionalAPIURLParametersNestedFieldsItem
-							);
-						}
-					}
-				);
-
-				url.searchParams.set('nestedFields', newNestedFields);
-			}
-			else {
-				url.searchParams.append(key, value);
-			}
-		});
-	}
-
-	const response = await fetch(url, {
-		headers: DEFAULT_FETCH_HEADERS,
-		method: 'GET',
-	});
-	const responseJSON = await response.json();
-
-	return {
-		data: responseJSON,
-		ok: response.ok,
-		status: response.status,
-	};
 }
 
 export function getCurrentItemUpdates(
