--- modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/utils/odata.js
@@ -1,8 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-export default function createOdataFilter(filters) {
-	return filters.map((filter) => `(${filter})`).join(' and ');
-}
